Silence -Wc++11-narrowing warnings
Clang treats these as errors by default, so this fixes the build with clang.

Example of one of the warnings:

/builddir/build/BUILD/pki-10.8.3/base/tps-client/src/processor/RA_Processor.cpp:3237:28:
error: non-constant-expression cannot be narrowed from type 'int' to
'BYTE' (aka 'unsigned char') in initializer list [-Wc++11-narrowing]

            BYTE nv[2] = { v, 0x01 };

+++ base/tps-client/src/processor/RA_Enroll_Processor.cpp
@@ -1769,7 +1769,7 @@ bool RA_Enroll_Processor::CheckAndUpgradeSymKeys(
 
 			// Assemble the Buffer with the version information
 			// The second byte is the key offset, which is always 1
-			BYTE nv[2] = { requiredV, 0x01 };
+			BYTE nv[2] = { static_cast<BYTE>(requiredV), 0x01 };
 			Buffer newVersion(nv, 2);
 
 			// GetKeyInfoData will return a buffer which is bytes 11,12 of
